requirements.txt
certifi==2019.11.28
chardet==3.0.4
cos-python-sdk-v5==1.7.8
dicttoxml==1.7.4
Django==1.11.7
idna==2.9
Pillow==7.0.0
pycrypto==2.6.1
pytz==2019.3
qcloud-python-sts==3.0.3
qcloudsms-py==0.1.4
redis==3.4.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8